Avalonia works and fixes for W/A

+++ AvaloniaLogReader/Commands/SettingsCommands/RemoveAllFilesCommand.cs
@@ -1,5 +1,4 @@
 ﻿using AvaloniaLogReader.Commands.BaseCommands;
-using System.Collections.Generic;
 using System.IO;
 
 namespace AvaloniaLogReader.Commands.SettingsCommands
@@ -14,18 +13,21 @@ namespace AvaloniaLogReader.Commands.SettingsCommands
 
         public override void Execute(object parameter)
         {
-            IEnumerable<string> contents = Directory.EnumerateFileSystemEntries(_path);
-
-            foreach (string path in contents)
+            if (Directory.Exists(_path))
             {
-                FileAttributes attributes = File.GetAttributes(path);
-                if (attributes.HasFlag(FileAttributes.Directory))
+                var contents = Directory.EnumerateFileSystemEntries(_path);
+
+                foreach (string path in contents)
                 {
-                    Directory.Delete(path, true);
-                }
-                else
-                {
-                    File.Delete(path);
+                    FileAttributes attributes = File.GetAttributes(path);
+                    if (attributes.HasFlag(FileAttributes.Directory))
+                    {
+                        Directory.Delete(path, true);
+                    }
+                    else
+                    {
+                        File.Delete(path);
+                    }
                 }
             }
         }
